+++ web/pages/nodes/_id/index.vue
@@ -3,8 +3,9 @@
     <b-row>
       <b-col md="12">
         <b-alert v-if="runningOperation !== null" variant="info" show>
-          <h5 class="alert-heading font-weight-bold">
-            {{ runningOperation.name }}
+          <h5 class="alert-heading font-weight-bold d-flex align-items-center">
+            <span>{{ runningOperation.name }}</span>
+            <b-spinner class="ml-auto" small variant="primary"></b-spinner>
           </h5>
           {{ runningOperation.log ? runningOperation.log.split('\n').slice(-1).pop() : '' }}
         </b-alert>